requirements.txt
numpy==2.1.2
pandas==2.2.3
scikit-learn==1.5.2
matplotlib==3.9.2
seaborn==0.13.2
jupyter==1.1.1
ipykernel==6.29.5
pyyaml==6.0.2
joblib==1.4.2
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
dvc==3.55.2
dvc-s3==3.2.0
mlflow==2.16.2
tensorflow-cpu==2.17.0
onnxruntime==1.19.2
optuna==3.6.1
pytest==8.3.3
requests==2.32.3
rich==13.9.2
typer==0.12.5
ruff==0.6.9


